site/requirements.txt
fastapi==0.115.12 
python-dotenv==1.0.1
asyncpg==0.30.0
uvicorn==0.34.0 
Jinja2==3.1.6 
pycryptodome==3.22.0 